UI form state: 3 sources for 3 donation levels with 3 options each. Values generated from: defaults, query parameters and user input; making sure these are mapped to the correct form field names submitted to the back end. Saving pre-refactor.
