+++ static/js/src/entry/business/BusinessForm.vue
@@ -282,9 +282,10 @@ import NativePay from '../../elements/NativePay.vue';
 import Choices from './Choices.vue';
 import updateStoreValue from '../../elements/mixins/updateStoreValue';
 import formStarter from '../../mixins/form/starter';
+// import addNumberCommas from '../../utils/addNumberCommas';
+import { US_STATES_SELECT_LIST } from '../../utils/formSelectListConstants';
 
 import {
-  US_STATES_SELECT_LIST,
   DEFAULT_STATE_SELECTED
 } from './constants';
